+++ fe2o3-amqp/src/link/delivery.rs
@@ -77,7 +77,7 @@ impl<T> Delivery<T> {
             Body::Value(AmqpValue(value)) => Ok(value),
             Body::Data(_) => Err(BodyError::IsData),
             Body::Sequence(_) => Err(BodyError::IsSequence),
-            Body::Nothing => Err(BodyError::IsNothing),
+            Body::Empty => Err(BodyError::IsEmpty),
         }
     }
 
@@ -88,7 +88,7 @@ impl<T> Delivery<T> {
             Body::Data(Data(data)) => Ok(data),
             Body::Value(_) => Err(BodyError::IsValue),
             Body::Sequence(_) => Err(BodyError::IsSequence),
-            Body::Nothing => Err(BodyError::IsNothing),
+            Body::Empty => Err(BodyError::IsEmpty),
         }
     }
 
@@ -99,7 +99,7 @@ impl<T> Delivery<T> {
             Body::Data(_) => Err(BodyError::IsData),
             Body::Sequence(AmqpSequence(sequence)) => Ok(sequence),
             Body::Value(_) => Err(BodyError::IsValue),
-            Body::Nothing => Err(BodyError::IsNothing),
+            Body::Empty => Err(BodyError::IsEmpty),
         }
     }
 
@@ -110,7 +110,7 @@ impl<T> Delivery<T> {
             Body::Value(AmqpValue(value)) => Ok(value),
             Body::Data(_) => Err(BodyError::IsData),
             Body::Sequence(_) => Err(BodyError::IsSequence),
-            Body::Nothing => Err(BodyError::IsNothing),
+            Body::Empty => Err(BodyError::IsEmpty),
         }
     }
 
@@ -121,7 +121,7 @@ impl<T> Delivery<T> {
             Body::Data(Data(data)) => Ok(data),
             Body::Value(_) => Err(BodyError::IsValue),
             Body::Sequence(_) => Err(BodyError::IsSequence),
-            Body::Nothing => Err(BodyError::IsNothing),
+            Body::Empty => Err(BodyError::IsEmpty),
         }
     }
 
@@ -132,7 +132,7 @@ impl<T> Delivery<T> {
             Body::Data(_) => Err(BodyError::IsData),
             Body::Sequence(AmqpSequence(sequence)) => Ok(sequence),
             Body::Value(_) => Err(BodyError::IsValue),
-            Body::Nothing => Err(BodyError::IsNothing),
+            Body::Empty => Err(BodyError::IsEmpty),
         }
     }
 }
@@ -143,7 +143,7 @@ impl<T: std::fmt::Display> std::fmt::Display for Delivery<T> {
             Body::Data(data) => write!(f, "{}", data),
             Body::Sequence(seq) => write!(f, "{}", seq),
             Body::Value(val) => write!(f, "{}", val),
-            Body::Nothing => write!(f, "Nothing"),
+            Body::Empty => write!(f, "Empty"),
         }
     }
 }